Move TrainStationRemovalArgs to RemoveTrainStation.h
