+++ src/OpenLoco/src/GameCommands/GameCommands.h
@@ -138,40 +138,6 @@ namespace OpenLoco::GameCommands
         return doCommand(T::command, regs);
     }
 
-    struct TrainStationRemovalArgs
-    {
-        static constexpr auto command = GameCommand::removeTrainStation;
-
-        TrainStationRemovalArgs() = default;
-        explicit TrainStationRemovalArgs(const registers& regs)
-            : pos(regs.ax, regs.cx, regs.di)
-            , rotation(regs.bh & 0x3)
-            , trackId(regs.dl & 0x3F)
-            , index(regs.dh & 0xF)
-            , type(regs.bp & 0xF)
-        {
-        }
-
-        World::Pos3 pos;
-        uint8_t rotation;
-        uint8_t trackId;
-        uint8_t index;
-        uint8_t type;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = pos.x;
-            regs.cx = pos.y;
-            regs.di = pos.z;
-            regs.bh = rotation;
-            regs.dl = trackId;
-            regs.dh = index;
-            regs.bp = type;
-            return regs;
-        }
-    };
-
     // Rename town
     inline void do_46(uint16_t cx, uint16_t ax, uint32_t edx, uint32_t ebp, uint32_t edi)
     {
